What I Learned
This project taught me about finding answers in documentation.  The project was completed very quickly but I wanted something that challenged me, so I set a goal of changing the font color depending on weather the out come was over or under budget.  It took a lot of digging and trial and error but eventually I was able to accomplish this task.